pyproject.toml
[project]
name = "parset-monitor"
version = "0.1.0"
description = "Intelligent product/listing monitoring MVP built with FastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiogram>=3.13.1",
  "apscheduler>=3.10.4",
  "beautifulsoup4>=4.12.3",
  "fastapi>=0.115.0",
  "httpx>=0.28.1",
  "jinja2>=3.1.4",
  "pydantic-settings>=2.6.1",
  "python-multipart>=0.0.18",
  "requests>=2.32.3",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
